reorganize images & imports, fix carousel behavior, add titles, fix browser errors
