+++ src/components/Hero/images.js
@@ -1,21 +1,21 @@
-import Camera from '../../images/camera.png';
-import Couch from '../../images/couch.png';
-import Desk from '../../images/desk.png';
-import FlowerPot from '../../images/flower-pot.png';
-import Kicks from '../../images/kicks.png';
-import PaintTube from '../../images/paint-tube.png';
-import Phone from '../../images/phone.png';
-import RecordPlayer from '../../images/record-player.png';
-import SprinkleArrow from '../../images/sprinkle-arrow.png';
-import SprinkleBurst from '../../images/sprinkle-burst.png';
-import SprinkleCircle from '../../images/sprinkle-circle.png';
-import SprinkleGrid from '../../images/sprinkle-grid.png';
-import SprinkleSlashes from '../../images/sprinkle-slashes.png';
-import SprinkleSparkle from '../../images/sprinkle-sparkle.png';
-import SprinkleSpring from '../../images/sprinkle-spring.png';
-import Stereo from '../../images/stereo.png';
-import Sticker from '../../images/sticker.png';
-import Stickers from '../../images/stickers.png';
+import Camera from '../../images/hero/camera.png';
+import Couch from '../../images/hero/couch.png';
+import Desk from '../../images/hero/desk.png';
+import FlowerPot from '../../images/hero/flower-pot.png';
+import Kicks from '../../images/hero/kicks.png';
+import PaintTube from '../../images/hero/paint-tube.png';
+import Phone from '../../images/hero/phone.png';
+import RecordPlayer from '../../images/hero/record-player.png';
+import SprinkleArrow from '../../images/hero/sprinkle-arrow.png';
+import SprinkleBurst from '../../images/hero/sprinkle-burst.png';
+import SprinkleCircle from '../../images/hero/sprinkle-circle.png';
+import SprinkleGrid from '../../images/hero/sprinkle-grid.png';
+import SprinkleSlashes from '../../images/hero/sprinkle-slashes.png';
+import SprinkleSparkle from '../../images/hero/sprinkle-sparkle.png';
+import SprinkleSpring from '../../images/hero/sprinkle-spring.png';
+import Stereo from '../../images/hero/stereo.png';
+import Sticker from '../../images/hero/sticker.png';
+import Stickers from '../../images/hero/stickers.png';
 
 const images = [
   {
