+++ src/images/hero/record-player.png
