+++ src/images/hero/sprinkle-arrow.png
