+++ src/images/hero/sprinkle-grid.png
