+++ src/images/hero/sprinkle-slashes.png
