+++ src/images/hero/sprinkle-sparkle.png
