+++ src/images/hero/stereo.png
