+++ src/images/hero/sticker.png
