+++ src/images/hero/stickers.png
